+++ migrations/1585388899167-add_default_file_storage_provider.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
 const CryptoJS = require('crypto-js');
-const { Configuration } = require('@schul-cloud/commons');
+const { Configuration } = require('@hpi-schul-cloud/commons');
 
 const { info, warning } = require('../src/logger');
 const { schoolModel: School } = require('../src/services/school/model');